Main menu, and menu that pops up for next level
I added the 'first' level called main menu, since it was the only way I could seem to get it to work. Be on that level when the game starts. Then I made it so when you exit the door some sparkles come up. 5 seconds after a menu saying congrats comes up and asks if you want to continue. I also did the one for the ai, but I don't know how to implement.